jdvjhqjhwjhdfbvbefkjvbaw;odhkadhjsgfjkmjxcbgvljqnsbglkjbsmnvbasknc vhjdhfjbadlicvkjsbdh fjjskjbcg psa hbckjdhcfjvhjfbpskvljhsbvddfjlbdpfkgbvkans  cvjbdfjgbbadslkvbhdbfbgjsbddljfbad hi(:

+++ java/src/Animation_manager.java
@@ -3,7 +3,7 @@
  */
 public interface Animation_manager
 {
-    public long get_animation_rate();
+    public int get_animation_rate();
     public Action create_animation_action(WorldModel world, int repeat_count);
     public void schedule_animation(WorldModel world, int repeat_count);
